changed loading from file

+++ lib/knapsack_solution_instance.hpp
@@ -2,19 +2,18 @@
 
 #include <iostream>
 #include <vector>
-
 #include "knapsack_problem_instance.hpp"
 
 class KnapsackSolutionInstance {
    public:
     KnapsackSolutionInstance(KnapsackProblemInstance pinstance) : pinstance_(pinstance) {};
-    uint cost();
+    uint64_t cost();
     void revert(); // Undo the last change
-    std::vector<std::pair<uint, double>> get_sorted_items_by_ratio() const;
-    void make_change(const std::vector<std::pair<uint, double>>& sorted_items);
+    std::vector<std::pair<uint64_t, double>> get_sorted_items_by_ratio() const;
+    void make_change(const std::vector<std::pair<uint64_t, double>>& sorted_items);
 
    private:
     KnapsackProblemInstance pinstance_;
-    std::vector<uint> csol_; // Current solution (list of item IDs)
-    std::vector<uint> prev_solution_; // Previous solution for revert
+    std::vector<uint64_t> csol_; // Current solution (list of item IDs)
+    std::vector<uint64_t> prev_solution_; // Previous solution for revert
 };
